Tuning_free_DNI/requirements.txt
absl-py==2.1.0
accelerate==0.31.0
antlr4-python3-runtime==4.9.3
av==12.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
diffusers==0.29.0
filelock==3.15.1
fsspec==2024.6.0
ftfy==6.2.0
grpcio==1.64.1
huggingface-hub==0.23.3
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
kornia==0.7.2
kornia_rs==0.1.3
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
pip==24.0
protobuf==4.25.3
psutil==5.9.8
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
setuptools==69.5.1
six==1.16.0
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.2
triton==2.3.0
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
xformers==0.0.26.post1
zipp==3.19.2
